move analytics to renderer
