+++ packages/insomnia/src/ui/__mocks__/analytics.ts
